refactor: moves news into feature folder
