+++ src/features/news/api/getNews.test.ts
@@ -1,6 +1,6 @@
 import { vi } from 'vitest';
-import * as fetcher from '../../utils/fetcher';
-import { fetchNews } from './newsFeed';
+import * as fetcher from '../../../utils/fetcher';
+import { getNews } from './getNews';
 
 const dummyApiUrl = 'foo-endpoint';
 
@@ -10,10 +10,10 @@ vi.mock('../../utils/config', () => ({
     },
 }));
 
-describe('fetchNews', () => {
+describe('getNews', () => {
     it('should call fetcher with the correct arguments', () => {
         const fetcherSpy = vi.spyOn(fetcher, 'default').mockResolvedValueOnce({});
-        void fetchNews();
+        void getNews();
         expect(fetcherSpy).toHaveBeenCalledWith('get', dummyApiUrl, 'data/v2/news/?lang=EN');
     });
 });
