+++ src/features/news/api/getNews.ts
@@ -1,0 +1,4 @@
+import config from '../../../utils/config';
+import fetcher from '../../../utils/fetcher';
+
+export const getNews = () => fetcher('get', config.apiUrl, 'data/v2/news/?lang=EN');
