+++ src/features/news/types/NewsArticle.ts
@@ -1,0 +1,9 @@
+export interface NewsArticle {
+    id: string;
+    publishedOn: string;
+    imageUrl: string;
+    title: string;
+    url: string;
+    body: string;
+    tags: string;
+}
