+++ src/features/news/types/NewsResult.ts
@@ -1,0 +1,5 @@
+import { NewsArticle } from './NewsArticle';
+
+export interface NewsResult {
+    Data: NewsArticle[];
+}
